rocm: Add new analysis to support new trace format
As a new plugin was added to generate CTF traces with different event
types and format, these new analysis and trace were added to support
this new type. Each API function now has its own event type, and some
fields are now passed as context fields.

As this change uses the tracecompass.analysis.callstack and the rocm
plugin depends on it, the callstack plugin was added as a dependency
to the trace server.

[Added] New ROCm trace type to support new plugin
[Added] New ROCm CallStack that displays API and GPU activity

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>